# swagger/notification_swagger.yaml
openapi: 3.0.0
info:
  title: User API
  description: API for user authentication, profile management, and MBTI settings
  version: 1.0.0
servers:
  - url: http://localhost:3000
paths:
  /notifications/{user_id}:
    get:
      summary: Get notifications
      description: All notifications for the authenticated user
      tags: 
        - Notifications
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: A list of notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    notification_id:
                      type: integer
                    user_id:
                      type: string
                    message:
                      type: string
                    status:
                      type: string
                    created_at:
                      type: string
                      format: date-time
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
      security:
        - sessionAuth: []
  /notifications:
    post:
      summary: Create a new notification
      description: Creates a new notification for a user
      tags: 
        - Notifications
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:
                  type: string
                message:
                  type: string
              required:
                - user_id
                - message
      responses:
        '201':
          description: Notification created
        '400':
          description: Invalid input
        '500':
          description: Internal server error
      security:
        - sessionAuth: []
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: connect.sid # express-session의 기본 쿠키 이름
security:
  - sessionAuth: []